+++ nais/prod-gcp/nais.yaml
@@ -20,6 +20,8 @@ spec:
   prometheus:
     enabled: true
     path: /tms-varselbjelle-api/metrics
+  secureLogs:
+    enabled: true
   replicas:
     min: 2
     max: 4
