+++ src/main/kotlin/no/nav/tms/varselbjelle/api/Application.kt
@@ -1,12 +1,19 @@
 package no.nav.tms.varselbjelle.api
 
+import io.ktor.client.*
+import io.ktor.client.engine.*
+import io.ktor.client.engine.apache.*
+import io.ktor.client.plugins.*
+import io.ktor.client.plugins.contentnegotiation.*
+import io.ktor.serialization.kotlinx.json.*
 import io.ktor.server.engine.embeddedServer
 import io.ktor.server.netty.Netty
 import no.nav.tms.token.support.azure.exchange.AzureServiceBuilder
 import no.nav.tms.varselbjelle.api.config.Environment
-import no.nav.tms.varselbjelle.api.config.HttpClientBuilder
 import no.nav.tms.varselbjelle.api.varsel.VarselService
 import no.nav.tms.varselbjelle.api.azure.AzureTokenFetcher
+import no.nav.tms.varselbjelle.api.config.HttpClientBuilder
+import no.nav.tms.varselbjelle.api.config.jsonConfig
 
 fun main() {
     val environment = Environment()
@@ -39,3 +46,4 @@ fun main() {
         )
     }.start(wait = true)
 }
+
