--- src/main/kotlin/no/nav/tms/varselbjelle/api/config/httpClient.kt
@@ -1,38 +1,0 @@
-package no.nav.tms.varselbjelle.api.config
-
-import io.ktor.client.HttpClient
-import io.ktor.client.call.body
-import io.ktor.client.engine.HttpClientEngine
-import io.ktor.client.engine.apache.Apache
-import io.ktor.client.plugins.HttpTimeout
-import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
-import io.ktor.client.request.header
-import io.ktor.client.request.request
-import io.ktor.client.request.url
-import io.ktor.http.HttpHeaders
-import io.ktor.http.HttpMethod
-import io.ktor.serialization.kotlinx.json.json
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
-import java.net.URL
-
-object HttpClientBuilder {
-
-    fun build(httpClientEngine: HttpClientEngine = Apache.create()): HttpClient {
-        return HttpClient(httpClientEngine) {
-            install(ContentNegotiation) {
-                json(jsonConfig())
-            }
-            install(HttpTimeout)
-        }
-    }
-}
-
-suspend inline fun <reified T> HttpClient.getForIdent(url: URL, fnr: String, accessToken: String): T = withContext(Dispatchers.IO) {
-    request {
-        url(url)
-        method = HttpMethod.Get
-        header("fodselsnummer", fnr)
-        header(HttpHeaders.Authorization, "Bearer $accessToken")
-    }
-}.body()
